修正 profilesテンプレートの名前変更、パスの修正
