+++ src/pages/profiles/school.js
@@ -7,28 +7,33 @@ import Grid from '@mui/material/Grid';
 import Container from '@mui/material/Container';
 import Button from '@mui/material/Button';
 import TextField from '@mui/material/TextField';
+import Select from '@mui/material/Select';
+import MenuItem from '@mui/material/MenuItem';
+import FormControl from '@mui/material/FormControl';
+import InputLabel from '@mui/material/InputLabel';
 import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography'
 import { Loading } from '../../components/Loading';
 import nookies from "nookies";
 import { useSnackbar } from 'notistack';
 
-
-const MAX_LINE_LENGTH_OF_ANSWER1 = 13;
-const MAX_LINE_LENGTH_OF_ANSWER2 = 13;
-const MAX_LINE_LENGTH_OF_ANSWER3 = 26;
-const MAX_LINE_COUNT = 3;
-
+const MAX_LINE_LENGTH_OF_ANSWER1 = 6;
+const MAX_LINE_LENGTH_OF_ANSWER2 = 12;
+const MAX_LINE_LENGTH_OF_ANSWER4 = 23;
+const MAX_LINE_LENGTH_OF_ANSWER5 = 15;
 
 export default function App () {
   const [body1, setBody1] = useState("");
   const [body2, setBody2] = useState("");
   const [body3, setBody3] = useState("");
-  const [imageUrl, setImageUrl] = useState("/templates/1.png");
+  const [body4, setBody4] = useState("");
+  const [body5, setBody5] = useState("");
+  const [imageUrl, setImageUrl] = useState("/templates/school.png");
 
   const [body1Error, setBody1Error] = useState("");
   const [body2Error, setBody2Error] = useState("");
-  const [body3Error, setBody3Error] = useState("");
+  const [body4Error, setBody4Error] = useState("");
+  const [body5Error, setBody5Error] = useState("");
 
   const { enqueueSnackbar } = useSnackbar();
 
@@ -40,7 +45,7 @@ export default function App () {
   const validateForm = () => {
     let isValid = true;
     if (!body1) {
-      enqueueSnackbar("ニックネームを入力してね", { variant: 'error' });
+      enqueueSnackbar("なまえを入力してね", { variant: 'error' });
       isValid = false;
     }
     if (!body2) {
@@ -48,16 +53,23 @@ export default function App () {
       isValid = false;
     }
     if (!body3) {
-      enqueueSnackbar("みんなにひとこと！を入力してね", { variant: 'error' });
+      enqueueSnackbar("推しCREDOを選択してね", { variant: 'error' });
+      isValid = false;
+    }
+    if (!body4) {
+      enqueueSnackbar("めざしたわけを入力してね", { variant: 'error' });
+      isValid = false;
+    }
+    if (!body5) {
+      enqueueSnackbar("がんばりたいことを入力してね", { variant: 'error' });
       isValid = false;
     }
     return isValid;
   };
 
-
   const handlePreview = async (e) => {
     e.preventDefault();
-    
+
   try {
     const cookies = nookies.get(null);
     const config = {
@@ -67,7 +79,7 @@ export default function App () {
       },
     };
 
-    const response = await axios.post(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/profiles/preview`, { answers: { body1, body2, body3 } }, config);
+    const response = await axios.post(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/profiles/school/preview`, { answers: { body1, body2, body3, body4, body5 } }, config);
     setImageUrl(response.data.url);
     } catch (error) {
     console.error(error);
@@ -90,7 +102,7 @@ export default function App () {
         },
       };
 
-      const response = await axios.post(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/profiles`, { answers: { body1, body2, body3 } }, config);
+      const response = await axios.post(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/profiles/school`, { answers: { body1, body2, body3, body4, body5 } }, config);
       
       if (response.status !== 200) {
         throw new Error('Request failed with status: ' + response.status);
@@ -145,7 +157,7 @@ export default function App () {
                   margin="normal"
                   fullWidth
                   id="answer1"
-                  label="ニックネームは？"
+                  label="なまえ"
                   name="answer1"
                   autoFocus
                   value={body1}
@@ -153,7 +165,7 @@ export default function App () {
                   helperText={body1Error}
                   onChange={(e) => {
                     if (e.target.value.length > MAX_LINE_LENGTH_OF_ANSWER1) {
-                      setBody1Error('13文字以内で入力してください。');
+                      setBody1Error('6文字以内で入力してください。');
                     } else {
                       setBody1Error('');
                       setBody1(e.target.value);
@@ -165,17 +177,54 @@ export default function App () {
                   margin="normal"
                   fullWidth
                   name="answer2"
-                  label="しゅみは？"
+                  label="しゅみ"
                   id="answer2"
                   value={body2}
                   error={body2Error !== ''}
                   helperText={body2Error}
                   onChange={(e) => {
                     if (e.target.value.length > MAX_LINE_LENGTH_OF_ANSWER2) {
-                      setBody2Error('13文字以内で入力してください。');
+                      setBody2Error('12文字以内で入力してください。');
                     } else {
                       setBody2Error('');
                       setBody2(e.target.value);
+                    }
+                  }}
+                />
+                <FormControl fullWidth>
+                  <InputLabel id="select-label">推しCREDO</InputLabel>
+                  <Select
+                    labelId="select-label"
+                    id="answer3"
+                    value={body3}
+                    label="推しCREDO"
+                    onChange={(e) => {
+                      setBody3(e.target.value);
+                    }}
+                  >
+                    <MenuItem value={"Be Open"}>Be Open</MenuItem>
+                    <MenuItem value={"Move Fast"}>Move Fast</MenuItem>
+                    <MenuItem value={"Give First"}>Give First</MenuItem>
+                    <MenuItem value={"Geek Out"}>Geek Out</MenuItem>
+                    <MenuItem value={"Take Ownership"}>Take Ownership</MenuItem>
+                  </Select>
+                </FormControl>
+                <TextField
+                  color="secondary"
+                  margin="normal"
+                  fullWidth
+                  name="answer4"
+                  label="めざしたわけ"
+                  id="answer4"
+                  value={body4}
+                  error={body4Error !== ''}
+                  helperText={body4Error}
+                  onChange={(e) => {
+                    if (e.target.value.length > MAX_LINE_LENGTH_OF_ANSWER4) {
+                      setBody4Error('23文字以内で入力してください。');
+                    } else {
+                      setBody4Error('');
+                      setBody4(e.target.value);
                     }
                   }}
                 />
@@ -183,21 +232,18 @@ export default function App () {
                   color="secondary"
                   margin="normal"
                   fullWidth
-                  name="answer3"
-                  label="みんなにひとこと！"
-                  id="answer3"
-                  value={body3}
-                  multiline
-                  rows={MAX_LINE_COUNT}
-                  error={body3Error !== ''}
-                  helperText={body3Error}
+                  name="answer5"
+                  label="がんばりたいこと"
+                  id="answer5"
+                  value={body5}
+                  error={body5Error !== ''}
+                  helperText={body5Error}
                   onChange={(e) => {
-                    const lines = e.target.value.split('\n');
-                    if (lines.length > MAX_LINE_COUNT || lines.some(line => line.length > MAX_LINE_LENGTH_OF_ANSWER3)) {
-                      setBody3Error('1行は26文字以内、改行は2回までとしてください。');
+                    if (e.target.value.length > MAX_LINE_LENGTH_OF_ANSWER5) {
+                      setBody5Error('15文字以内で入力してください。');
                     } else {
-                      setBody3Error('');
-                      setBody3(e.target.value);
+                      setBody5Error('');
+                      setBody5(e.target.value);
                     }
                   }}
                 />
